pyproject.toml
[tool.poetry]
name = "example-qa-assessment"
version = "0.1.0"
description = ""
authors = ["Carlos Kidman <carlos@qap.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.0.1"
requests = "^2.27.1"
pydantic = "^1.9.0"
pyleniumio = "^1.14.1"
pytest-html = "^3.1.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
